pyproject.toml
[tool.poetry]
name = "the tractor guy"
version = "0.0.1.0rc"
description = ""
authors = ["Inzamul Hoque <inzamol@gmail.com>"]
license = "RBS"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
djangorestframework = "^3.15.2"
markdown = "^3.7"
django-filter = "^24.3"
drf-yasg = "^1.21.8"
rest-framework-simplejwt = "^0.0.2"
djangorestframework-simplejwt = "5.3.0"
django-migration-rollback = "^1.0.5"
drf-standardized-errors = "^0.14.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
